move fingerprint method to training module
